etl/requirements.txt
numpy==2.4.2
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
